Move existing controllers
