+++ app/controllers/bill-runs.controller.js
@@ -7,8 +7,8 @@
 
 const Boom = require('@hapi/boom')
 
-const CreateBillRunValidator = require('../../validators/bill-runs/create-bill-run.validator.js')
-const NewBillingBatchService = require('../../services/billing/new-billing-batch.service.js')
+const CreateBillRunValidator = require('../validators/bill-runs/create-bill-run.validator.js')
+const NewBillingBatchService = require('../services/billing/new-billing-batch.service.js')
 
 async function create (request, h) {
   const validatedData = CreateBillRunValidator.go(request.payload)
